refactor edit into app folder
