+++ components/form-builder/app/edit/ElementPanel.tsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from "react";
 import { useTranslation } from "next-i18next";
 
-import { FormElementWithIndex } from "../types";
-import { useTemplateStore, useModalStore } from "../store";
+import { FormElementWithIndex } from "../../types";
+import { useTemplateStore, useModalStore } from "../../store";
 import { PanelActions, ModalButton, ModalForm, PanelBody } from "./index";
-import { Button } from "../shared";
+import { Button } from "../../shared";
 
 export const ElementPanel = ({ item }: { item: FormElementWithIndex }) => {
   const { t } = useTranslation("form-builder");
