+++ components/form-builder/app/edit/ModalForm.tsx
@@ -4,9 +4,9 @@ import PropTypes from "prop-types";
 import { useTranslation } from "next-i18next";
 import { FormElementTypes, ElementProperties } from "@lib/types";
 
-import { FormElementWithIndex, LocalizedElementProperties } from "../types";
-import { Checkbox, Input, TextArea } from "../shared";
-import { useTemplateStore } from "../store";
+import { FormElementWithIndex, LocalizedElementProperties } from "../../types";
+import { Checkbox, Input, TextArea } from "../../shared";
+import { useTemplateStore } from "../../store";
 
 const FormLabel = styled.label`
   font-weight: 700;
