+++ components/form-builder/app/edit/PanelActions.tsx
@@ -2,11 +2,11 @@ import React, { useState, useCallback, useEffect, useRef, KeyboardEvent } from "
 import PropTypes from "prop-types";
 import { useTranslation } from "next-i18next";
 
-import { Button } from "../shared";
-import { ChevronUp, ChevronDown, Close, Duplicate, ThreeDotsIcon } from "../icons";
+import { Button } from "../../shared";
+import { ChevronUp, ChevronDown, Close, Duplicate, ThreeDotsIcon } from "../../icons";
 import { Modal } from "./Modal";
-import { FormElementWithIndex } from "../types";
-import { useTemplateStore } from "../store";
+import { FormElementWithIndex } from "../../types";
+import { useTemplateStore } from "../../store";
 
 export const PanelActions = ({
   item,
