+++ components/form-builder/app/edit/SelectedElement.tsx
@@ -2,11 +2,11 @@ import React, { useCallback } from "react";
 import { FormElementTypes, HTMLTextInputTypeAttribute } from "@lib/types";
 import { useTranslation } from "next-i18next";
 
-import { CheckBoxEmptyIcon, RadioEmptyIcon } from "../icons";
-import { ShortAnswer, Options, RichText } from "../elements";
-import { ElementOption, FormElementWithIndex } from "../types";
-import { useElementOptions } from "../hooks";
-import { useTemplateStore } from "../store";
+import { CheckBoxEmptyIcon, RadioEmptyIcon } from "../../icons";
+import { ShortAnswer, Options, RichText } from "../../elements";
+import { ElementOption, FormElementWithIndex } from "../../types";
+import { useElementOptions } from "../../hooks";
+import { useTemplateStore } from "../../store";
 
 export const SelectedElement = ({
   selected,
